feat: refactor header component structure and localization for version two
