--- src/components/molecules/Header/icons/hamburger/styles.css
@@ -1,23 +1,0 @@
-#hamburger-wrapper,
-#top-bread,
-#meat,
-#lettuce,
-#bottom-bread {
-  transition: all 0.2s ease-in-out;
-}
-
-#hamburger-opened #top-bread {
-  transform: translateY(-3px);
-}
-
-#hamburger-opened #meat {
-  transform: translateY(-1px);
-}
-
-#hamburger-opened #lettuce {
-  transform: translateY(1px);
-}
-
-#hamburger-opened #bottom-bread {
-  transform: translateY(4px);
-}
