+++ src/tests/home.spec.ts
@@ -1,7 +1,6 @@
 import { test, expect } from "@playwright/test";
 
 import enTranslations from "../i18n/locales/en.json";
-import ptTranslations from "../i18n/locales/pt.json";
 
 test.describe("Locales", () => {
   test.describe("US locale user", () => {
@@ -21,31 +20,7 @@ test.describe("Locales", () => {
       await page.goto("/");
 
       await expect(page.locator("#my-title")).toHaveText(
-        enTranslations.presentation.title,
-      );
-    });
-  });
-
-  test.describe("Brazil locale", () => {
-    test.use({
-      locale: "pt-BR",
-    });
-
-    test("if the HTML lang attribute is in brazilian portuguese", async ({
-      page,
-    }) => {
-      await page.goto("/");
-
-      await expect(page.locator("html")).toHaveAttribute("lang", "pt");
-    });
-
-    test("if the page text is in brazilian portuguese for Brazil located users", async ({
-      page,
-    }) => {
-      await page.goto("/");
-
-      await expect(page.locator("#my-title")).toHaveText(
-        ptTranslations.presentation.title,
+        enTranslations.home.title,
       );
     });
   });
@@ -67,7 +42,7 @@ test.describe("Locales", () => {
       await page.goto("/");
 
       await expect(page.locator("#my-title")).toHaveText(
-        enTranslations.presentation.title,
+        enTranslations.home.title,
       );
     });
   });
@@ -83,7 +58,7 @@ test.describe("Page elements", () => {
   test("if the page has every content section visible", async ({ page }) => {
     await page.goto("/");
 
-    await expect(page.locator("#presentation")).toBeVisible();
+    await expect(page.locator("#home")).toBeVisible();
     await expect(page.locator("#about-me")).toBeVisible();
     await expect(page.locator("#articles")).toBeVisible();
     await expect(page.locator("#references")).toBeVisible();
@@ -102,11 +77,9 @@ test.describe("Page elements", () => {
     await expect(page.locator("#footer")).toBeVisible();
   });
 
-  test("if the page has the presentation section i the viewport", async ({
-    page,
-  }) => {
+  test("if the page has the home section i the viewport", async ({ page }) => {
     await page.goto("/");
 
-    await expect(page.locator("#presentation")).toBeInViewport();
+    await expect(page.locator("#home")).toBeInViewport();
   });
 });
